% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_runtime.R
\name{calc_runtime}
\alias{calc_runtime}
\title{Calculate the running time for serial measurement data}
\usage{
calc_runtime(
  df,
  ...,
  .id = encntr_id,
  .event = event,
  .dt_tm = event_datetime,
  .units = "hours"
)
}
\arguments{
\item{df}{A data frame}

\item{...}{Optional columns to group by}

\item{.id}{Patient identifier column, defaults to encntr_id}

\item{.event}{Event column, defaults to event}

\item{.dt_tm}{Date/time column, defaults to event_datetime}

\item{.units}{A string specifying the time units to use in calculations,
default is "hours"}
}
\value{
A data frame
}
\description{
\code{calc_runtime} calculates the duration of time at current value and
total run time
}
\details{
This function takes a data frame with serial measurement data and produces a
data frame with the with the duration of time at each rate and cumulative run
time.
}
