% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sofa.R
\name{calc_sofa}
\alias{calc_sofa}
\title{Calculate SOFA score}
\usage{
calc_sofa(
  df,
  .id = encntr_id,
  .pa02 = pao2,
  .fio2 = fio2,
  .intubated = intubated,
  .platelet = platelet,
  .gcs = gcs,
  .bili = bili,
  .map = map,
  .scr = scr,
  .uop = uop,
  .dopamine = dopamine,
  .dobutamine = dobutamine,
  .epinephrine = epinephrine,
  .norepinephrine = norepinephrine
)
}
\arguments{
\item{df}{data_frame}

\item{...}{optional named arguments with column names}
}
\value{
data_frame
}
\description{
This takes a data frame with columns for each parameter need to calculate the
SOFA score. The columns containing each of these parameters can optionally be
passed as named arguments if they differ from the default column names.
}
\references{
Sequential Organ Failure Assessment (SOFA) Score: https://www.mdcalc.com/calc/691/sequential-organ-failure-assessment-sofa-score
}
