% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat_encounters.R
\name{concat_encounters}
\alias{concat_encounters}
\title{Make concatenated list of encounters}
\usage{
concat_encounters(encounters, num.split = 900)
}
\arguments{
\item{encounters}{A character vector with the encounters}

\item{num.split}{An optional numeric indicating the number to split on,
default is 900}
}
\value{
A list
}
\description{
\code{concat_encounters} takes a vector of encounters and produces a
concatenated list for use in EDW
}
\details{
This function takes a vector of encounters and concatenates them together,
separated by a semi-colon. The list is then split into the desired group
size. The resulting string can be used in an EDW query prompt.
}
