% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_change.R
\name{count_rowsback}
\alias{count_rowsback}
\title{Count the number of rows to go back in data frame}
\usage{
count_rowsback(x, back = 2)
}
\arguments{
\item{x}{= vector of type POSIXct}

\item{back}{= integer indicating the number of days back to include}
}
\value{
integer vector
}
\description{
Takes a vector of POSIXct and counts the number of rows which would fall
within the specified time frame. Typically called from
\code{mutate} and the results are passed on to
\code{rollapplyr}.
}
\keyword{internal}
