% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_runtime.R
\name{drip_runtime}
\alias{drip_runtime}
\title{Calculate the running time for continuous infusion data}
\usage{
drip_runtime(df, .grp_var, ..., drip_off = 12, no_doc = 24, units = "hours")
}
\arguments{
\item{df}{A data frame}

\item{.grp_var}{column names to group by, wrapped by dplyr::vars}

\item{...}{optional named arguments with column names; optional arguments
include: med, med_datetime, rate, and rate_unit; if not specified,
medication, med_datetime, rate, and rate_unit are used for column names}

\item{drip_off}{An optional numeric indicating the number of hours a
medication infusion should be off to count as a new infusion, defaults to
12 hours}

\item{no_doc}{An optional numeric indicating the number of hours without
documentation which will be used to indicate a drip has ended, defaults to
24 hours}

\item{units}{An optional character string specifying the time units to use in
calculations, default is hours}
}
\value{
A data frame
}
\description{
\code{drip_runtime} calculates the duration of time at current value and
total run time
}
\details{
This function takes a data frame with continuous infusion data and produces a
data frame with the with the duration of time at each rate and cumulative run
time.

This could be used to then calculate the AUC or to summarize the continuous
data.

The data will be grouped into distinct sets of infusions, for patients who
may have been restarted on the drip one or more times. Use the
\code{drip_off} argument to modify the criteria for determining distinct
infusions.
}
