% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get data from csv files}
\usage{
get_data(path, pattern, col_types = NULL, tz = "US/Central")
}
\arguments{
\item{path}{a character vector of full path names}

\item{pattern}{a regular expression.  Only file names which match the regular
expression will be returned.}

\item{col_types}{one of NULL (default), a cols() specification, or a string}

\item{tz}{time zone}
}
\value{
A tibble
}
\description{
Read data from one or more csv files into a single data frame (tibble)
}
