% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_xlsx_data}
\alias{get_xlsx_data}
\title{Get data from xlsx files}
\usage{
get_xlsx_data(
  path,
  pattern,
  sheet = NULL,
  col_names = TRUE,
  col_types = NULL,
  skip = 0
)
}
\arguments{
\item{path}{a character vector of full path names}

\item{pattern}{a regular expression.  Only file names which match the regular
expression will be returned.}

\item{sheet}{sheet to read, either a string (the name of a sheet), or an integer (the position of the sheet).}

\item{col_names}{\code{TRUE} to use the first row as column names,
\code{FALSE} to get default names, or a character vector giving a name for
each column}

\item{col_types}{Either \code{NULL} to guess all from the spreadsheet or a
character vector containing one entry per column from these options:
"skip", "guess", "logical", "numeric", "date", "text" or "list".}

\item{skip}{Minimum number of rows to skip before reading anything, be it
column names or data}
}
\value{
A tibble
}
\description{
Read data from one or more Excel files into a single data frame (tibble)
}
