% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbohelpr.R
\docType{package}
\name{mbohelpr}
\alias{mbohelpr}
\title{Helper Functions for Data from MBO}
\description{
\pkg{mbohelpr} provides functions to facilitate working with data extracted
from MBO
}
\details{
This package provides standardized functions for reading and transforming
data which was extracted from Millennium Business Objects (MBO)
}
