% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_runtime.R
\name{med_runtime}
\alias{med_runtime}
\title{Calculate the running time for intermittent medication data}
\usage{
med_runtime(
  df,
  ...,
  .id = encntr_id,
  .med = medication,
  .dt_tm = med_datetime,
  .dose = dose,
  .med_off = 36,
  .no_doc = 24,
  .units = "hours"
)
}
\arguments{
\item{df}{A data frame}

\item{...}{Optional columns to group by}

\item{.id}{Patient identifier column, defaults to encntr_id}

\item{.med}{Medication column, defaults to medication}

\item{.dt_tm}{Date/time column, defaults to med_datetime}

\item{.dose}{Dose column, defaults to dose}

\item{.med_off}{Number of hours between medication doses which will be
counted as a new course, defaults to 36 hours}

\item{.no_doc}{Number of hours without documentation which will be used to
indicate a course has ended, defaults to 24 hours}

\item{.units}{A string specifying the time units to use in calculations,
default is "hours"}
}
\value{
A data frame
}
\description{
\code{med_runtime} calculates the duration of time at current value and
total run time
}
\details{
This function takes a data frame with intermittent medication data and produces a
data frame with the with the duration of time at each rate and cumulative run
time.

This could be used to then calculate the AUC or to summarize the medication course data.

The data will be grouped into distinct courses of medications, for patients who
may have been restarted on the medication one or more times. Use the
\code{med_off} argument to modify the criteria for determining distinct
courses.
}
