% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_data_path.R
\name{set_data_path}
\alias{set_data_path}
\title{Set the path to data files}
\usage{
set_data_path(folder, project = "")
}
\arguments{
\item{folder}{A string, the name of the top-level directory containing the
project}

\item{project}{An optional string, the name of the sub-folder containing an
individual research project}
}
\value{
A string
}
\description{
\code{set_data_path} sets the path to data files based on the OS
}
\details{
This function takes the name of a project and folder and sets the path to the
data based on the OS being used and availability of the network drive.
}
