% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_data.R
\name{summarize_data}
\alias{summarize_data}
\title{Summarize serial measurement data}
\usage{
summarize_data(
  df,
  ...,
  .id = encntr_id,
  .event = event,
  .dt_tm = event_datetime,
  .result = result
)
}
\arguments{
\item{df}{A data frame with continuous data}

\item{...}{Optional columns to group by}

\item{.id}{Patient identifier column, defaults to encntr_id}

\item{.event}{Event column, defaults to event}

\item{.dt_tm}{Date/time column, defaults to event_datetime}

\item{.result}{Result column, defaults to result}
}
\value{
tibble
}
\description{
\code{summarize_data} calculates summary data for serial measurements.
}
\details{
This function takes a data frame with serial measurement data which has been
transformed by \code{calc_runtime} and produces a data frame with summary
data for each patient.
}
