% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_data.R
\name{summarize_drips}
\alias{summarize_drips}
\title{Summary calculations for continuous infusion data}
\usage{
summarize_drips(
  df,
  ...,
  .id = encntr_id,
  .med = medication,
  .rate = rate,
  .units = "hours"
)
}
\arguments{
\item{df}{A data frame with continuous data}

\item{...}{Optional columns to group by}

\item{.id}{Patient identifier column, defaults to encntr_id}

\item{.med}{Medication column, defaults to medication}

\item{.rate}{Infusion rate column, defaults to rate}

\item{.units}{A string specifying the time units to use in calculations,
default is "hours"}
}
\value{
A data frame
}
\description{
\code{summarize_drips} summarizes continuous infusion data
}
\details{
This function takes a data frame with continuous infusion data which contains
the running time at each rate and produces a data frame with summary data for
each patient. The summary calculations include: first rate, last rate,
minimum rate, maximum rate, AUC, time-weighted average rate, total infusion
duration, total infusion running time, and cumulative dose.
}
