% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_data.R
\name{summarize_home_meds}
\alias{summarize_home_meds}
\title{Summarize home meds}
\usage{
summarize_home_meds(x, ..., ref, pts = NULL, home = TRUE)
}
\arguments{
\item{x}{data frame}

\item{...}{columns}

\item{ref}{a reference}

\item{pts}{patients}

\item{home}{home meds vs discharge prescriptions}
}
\description{
Summarize home meds
}
\details{
The data frame passed to \code{ref} should contain three character
  columns: name, type, and group. The name column should contain either
  generic medication names or medication classes. The type column should
  specify whether the value in name is a "class" or "med". The group column
  should specify whether the medication is a continous ("cont") or scheduled
  ("sched") medication.
}
