% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_runtime.R
\name{calc_runtime}
\alias{calc_runtime}
\title{Calculate the running time for serial measurement data}
\usage{
calc_runtime(df, ..., .grp_var, units = "hours")
}
\arguments{
\item{df}{A data frame}

\item{...}{optional named arguments with column names; optional arguments
are: id, event, and event_datetime; if not specified, encntr_id, event, and
event_datetime are used for column names, respectively}

\item{.grp_var}{column names to group by, uses \code{tidy_select}}

\item{units}{An optional character string specifying the time units to use in
calculations, default is hours}
}
\value{
A data frame
}
\description{
\code{calc_runtime} calculates the duration of time at current value and
total run time
}
\details{
This function takes a data frame with serial measurement data and produces a
data frame with the with the duration of time at each rate and cumulative run
time.
}
