% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_runtime.R
\name{drip_runtime}
\alias{drip_runtime}
\title{Calculate the running time for continuous infusion data}
\usage{
drip_runtime(
  df,
  ...,
  .id = encntr_id,
  .med = medication,
  .dt_tm = med_datetime,
  .rate = rate,
  .rate_unit = rate_unit,
  .drip_off = 12,
  .no_doc = 24,
  .units = "hours"
)
}
\arguments{
\item{df}{A data frame}

\item{...}{Optional columns to group by}

\item{.id}{Patient identifier column, defaults to encntr_id}

\item{.med}{Medication column, defaults to medication}

\item{.dt_tm}{Date/time column, defaults to med_datetime}

\item{.rate}{Infusion rate column, defaults to rate}

\item{.rate_unit}{Infusion rate units column, defaults to rate_unit}

\item{.drip_off}{Number of hours a medication infusion should be off to count
as a new infusion, defaults to 12 hours}

\item{.no_doc}{Number of hours without documentation which will be used to
indicate a drip has ended, defaults to 24 hours}

\item{.units}{A string specifying the time units to use in calculations,
default is "hours"}
}
\value{
A data frame
}
\description{
\code{drip_runtime} calculates the duration of time at current value and
total run time
}
\details{
This function takes a data frame with continuous infusion data and produces a
data frame with the with the duration of time at each rate and cumulative run
time.

This could be used to then calculate the AUC or to summarize the continuous
data.

The data will be grouped into distinct sets of infusions, for patients who
may have been restarted on the drip one or more times. Use the
\code{.drip_off} argument to modify the criteria for determining distinct
infusions.
}
