% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_change.R
\name{lab_change}
\alias{lab_change}
\title{Determine if a lab changed by a set amount within a specific time frame}
\usage{
lab_change(
  df,
  change_by,
  FUN,
  back = 2,
  .id = encntr_id,
  .dt_tm = lab_datetime,
  .result = result_val
)
}
\arguments{
\item{df}{A data frame with lab data}

\item{change_by}{A numeric indicating the threshold for lab changes}

\item{FUN}{A function for \code{rollapplyr}, most commonly max or
min}

\item{back}{An optional numeric specifying the number of days back to go.
Defaults to 2 days.}

\item{.id}{Patient identifier column, defaults to encntr_id}

\item{.dt_tm}{Date/time column, defaults to lab_datetime}

\item{.result}{Lab result, defaults to result_val}
}
\value{
A data frame
}
\description{
\code{lab_change} checks for changes in lab values
}
\details{
This function takes a data frame with lab data for a single lab and checks
whether the lab changes by a certain amount within a given period of time.

For FUN, use max when looking for a decrease in lab value, and min when
looking for an increase in lab value.
}
\examples{
# checks for a >= 2 decrease in the hemoglobin value within the past 2 days
x <- tidy_data(labs)

print(head(
  lab_change(x, "hgb", -2, max, back = 2)
))

}
