% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_runtime.R
\name{med_runtime}
\alias{med_runtime}
\title{Calculate the running time for intermittent medication data}
\usage{
med_runtime(df, ..., .grp_var, med_off = 24, no_doc = 24, units = "hours")
}
\arguments{
\item{df}{A data frame}

\item{...}{optional named arguments with column names; optional arguments
are: id, med, med_datetime, and dose; if not specified, encntr_id, medication,
med_datetime, and dose are used for column names, respectively}

\item{.grp_var}{column names to group by, uses \code{tidy_select}}

\item{med_off}{An optional numeric indicating the number of hours between
medication doses which will be counted as a new course, defaults to 36
hours}

\item{no_doc}{An optional numeric indicating the number of hours without
documentation which will be used to indicate a course has ended, defaults
to 24 hours}

\item{units}{An optional character string specifying the time units to use in
calculations, default is hours}
}
\value{
A data frame
}
\description{
\code{med_runtime} calculates the duration of time at current value and
total run time
}
\details{
This function takes a data frame with intermittent medication data and produces a
data frame with the with the duration of time at each rate and cumulative run
time.

This could be used to then calculate the AUC or to summarize the medication course data.

The data will be grouped into distinct courses of medications, for patients who
may have been restarted on the medication one or more times. Use the
\code{med_off} argument to modify the criteria for determining distinct
courses.
}
