% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_data.R
\name{summarize_data}
\alias{summarize_data}
\title{Summarize serial measurement data}
\usage{
summarize_data(df, ...)
}
\arguments{
\item{df}{A data frame with continuous data}

\item{...}{optional named arguments with column names}
}
\value{
tibble
}
\description{
\code{summarize_data} calculates summary data for serial measurements.
}
\details{
This function takes a data frame with serial measurement data which has been
transformed by \code{calc_runtime} and produces a data frame with summary
data for each patient.
}
